requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==2.1.15
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.2
kombu==5.4.2
prompt_toolkit==3.0.48
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
